fix(audit): retention day columns as integer, not bigint
The audit migration declared audit_retention_days and trash_retention_days
as int8. Kysely surfaces int8 reads as string (bigint can exceed
Number.MAX_SAFE_INTEGER), so app code that treats them as number broke
TypeScript compilation in three places after the recent db.d.ts regen.

- New migration ALTERs both columns to integer for existing DBs.
- Original audit migration updated so fresh installs get integer directly.
- db.d.ts updated to number | null for both columns.

+++ apps/server/src/database/migrations/20260228T223532-audit.ts
@@ -36,12 +36,12 @@ export async function up(db: Kysely<any>): Promise<void> {
   // add new workspace columns
   await db.schema
     .alterTable('workspaces')
-    .addColumn('audit_retention_days', 'int8', (col) => col)
+    .addColumn('audit_retention_days', 'integer', (col) => col)
     .execute();
 
   await db.schema
     .alterTable('workspaces')
-    .addColumn('trash_retention_days', 'int8', (col) => col)
+    .addColumn('trash_retention_days', 'integer', (col) => col)
     .execute();
 }
